requirements.txt
django==4.2.6
djangorestframework==3.14.0
grpcio==1.58.0
grpcio-tools==1.58.0
pdfplumber==0.10.2
python-docx==0.8.11
Pillow==10.0.1
vosk==0.3.45
ffmpeg-python==0.2.0
